+++ custom_components/smartthings_web/button.py
@@ -8,7 +8,7 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SmartThingsWebConfigEntry
-from .bridge_client import BridgeClientError
+from .bridge_client import BridgeClientError, bridge_error_message
 from .entity import SmartThingsWebDeviceEntity
 from .models import (
     BridgeControl,
@@ -89,4 +89,4 @@ class SmartThingsWebRefreshButton(SmartThingsWebDeviceEntity, ButtonEntity):
                 arguments=[],
             )
         except BridgeClientError as err:
-            raise HomeAssistantError("SmartThings Web did not confirm the button action") from err
+            raise HomeAssistantError(bridge_error_message("button command", err)) from err