+++ custom_components/smartthings_web/cover.py
@@ -10,7 +10,7 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SmartThingsWebConfigEntry
-from .bridge_client import BridgeClientError
+from .bridge_client import BridgeClientError, bridge_error_message
 from .entity import SmartThingsWebDeviceEntity
 from .models import (
     BridgeControl,
@@ -131,7 +131,7 @@ class SmartThingsWebCover(SmartThingsWebDeviceEntity, CoverEntity):
                 arguments=arguments,
             )
         except BridgeClientError as err:
-            raise HomeAssistantError("SmartThings Web did not confirm cover state") from err
+            raise HomeAssistantError(bridge_error_message("cover command", err)) from err
 
 
 def _find_control(controls: list[BridgeControl], *commands: str) -> BridgeControl | None: