+++ custom_components/smartthings_web/select.py
@@ -8,7 +8,7 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import SmartThingsWebConfigEntry
-from .bridge_client import BridgeClientError
+from .bridge_client import BridgeClientError, bridge_error_message
 from .entity import SmartThingsWebDeviceEntity
 from .models import (
     BridgeControl,
@@ -98,4 +98,4 @@ class SmartThingsWebSelect(SmartThingsWebDeviceEntity, SelectEntity):
                 arguments=[option],
             )
         except BridgeClientError as err:
-            raise HomeAssistantError("SmartThings Web did not confirm select state") from err
+            raise HomeAssistantError(bridge_error_message("select command", err)) from err